Update create_network_topology_file.py
Added these changes:
1) Round tosegment and segid values from float to int.
The values of the tosegments were written in float numbers
in the river_shape file. This creates problems in finding
segidment ids drain into downstream

2) Added the condition whether the river_outlet_id was not
found. If the the ID of the most downstream segment was
not defined in the controlFile, for instance in the
application which we have multiple outlets, we don't have
to set the river_down_seg_id to zero.